Introducing `Hanami::Application::Routing` namespace
